requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
idna==2.7
PyJWT==1.7.1
pytz==2019.3
requests==2.20.1
sqlparse==0.3.0
urllib3==1.24.3
